backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
deepface==0.0.89
numpy==1.26.3
pillow==10.2.0
opencv-python-headless==4.9.0.80
tf-keras==2.16.0
aiofiles==23.2.1
